This is a Java class that represents the doctor information, contains the ID, name, basic information, and provides a constructor and print method.
When writing this program, my goal was to create a class that would generally represent basic information about doctors. I first defined three attributes: id (for the doctor's number), title (for the doctor's name or title, such as "Dr. Wang"), and basicInfo (to store basic information about the doctor, such as "Doctor" or professional background).

Then, I designed two constructors for this class:

1. Default constructor: This constructor takes no arguments and initializes all properties to null or default values (such as 0 or an empty string) so that objects can be created without providing data.
2. Parameterized constructor: This constructor allows me to assign values directly when creating the object, such as specifying the ID, name, and basic information of the doctor, so that the data is set up at initialization time.
Finally, I added a print() method to print the values of these properties to the console. This way, you can easily view the doctor's information by creating an instance of the class and calling this method.

This program is a simple example of object-oriented programming, showing how to encapsulate data with classes and objects and manipulate them.