GeneralPractitioner is a subclass that inherits HealthProfessional, adding the address attribute and overwriting the print method.
When writing this program, my goal was to create a subclass that inherited from HealthProfessional, GeneralPractitioner, and add attributes specific to the general practitioner, such as an address. Here are my specific writing steps:

1. Clarify the succession relationship:
I first decided to let GeneralPractitioner inherit HealthProfessional so that the generic attributes of ID, name, and base information from the parent class can be reused directly without redefinition.

2. Create a class declaration:
I wrote public class GeneralPractitioner extends HealthProfessional, making it clear that GeneralPractitioner is a subclass of HealthProfessional.

3. Define unique attributes:
To distinguish GeneralPractitioner from other possible types of doctors, I added a private String address attribute that represents the general practitioner's address. This is unique to general practitioners and is therefore defined in the subcategory.

4. Write constructors:

4.1 Default Constructor: I implemented a parameterless constructor. In this function, I call super() to initialize the default properties of the superclass and set the address to the empty string "".
4.2 Constructor with arguments: To make it easy to create an object that contains all properties, I wrote a constructor with arguments. this function calls super(id, name, basicInfo) to initialize the properties of the parent class and sets the address of the subclass using this.address = address.
5. Rewrite the print method:
I want the GeneralPractitioner message output to include address information, so I override the parent class's print() method.

5.1 In the method, I first called the print logic of the parent class through super.print() to output the information of the common attribute.
5.2 Then adds a line of code with System.out.println("Address: "+ this.address); To print address information specific to the subclass.
6. Verification function:
To ensure that the class functions properly, I create GeneralPractitioner objects in the main program and test them using the default constructor and the parameterized constructor respectively. They then call their print() method to verify that the information is fully printed.