The AnotherProfessional class inherits from HealthProfessional and adds a new attribute, profession, that represents a profession. At the same time, the initialization and information output of the class are realized through the constructor and the overridden print() method.
When writing this program, my goal was to create an AnotherProfessional class that represents other medical professionals. To avoid duplication of code, I made this class inherit from HealthProfessional so that common properties and methods can be reused directly.

1. Declaration of succession:
I use extends HealthProfessional in my class declaration to make it clear that this is a subclass that inherits from its parent class.

2. New properties:
I've defined a private String profession to represent a profession, which is unique to the AnotherProfessional class.

3. Write constructors:

3.1 Default constructor:
In the no-argument constructor, I call the default constructor of the superclass super() to initialize the properties of the superclass while setting profession to the default value "".
3.2 Constructors with parameters:
I wrote a constructor with ID, name, basic information, and specialty. In this function, the attribute of the parent class is initialized by super(id, name, basicInfo), and the specialty attribute of the subclass is set by this.profession = profession.
4. Rewrite the print method:
To include specialized information in the output, I overwrote the parent class's print() method.

4.1 First call the print logic of the superclass through super.print() to output the generic properties.
4.2 Then use System.out.println("profession: "+ this.profession); Print professional properties.
5. Verification:
In the main program, I create several AnotherProfessional objects, test them with the default constructor and the parameterizing constructor, and call the print() method to check that the output is correct.